+++ MANIFEST.in
@@ -1,0 +1,5 @@
+include README.md
+include LICENSE
+include requirements.txt
+recursive-include tests *
+recursive-include example *
