+++ PySubtitle.egg-info/SOURCES.txt
@@ -1,0 +1,21 @@
+LICENSE
+MANIFEST.in
+README.md
+requirements.txt
+setup.py
+PySubtitle/__init__.py
+PySubtitle/audio_extraction.py
+PySubtitle/speech_recognition.py
+PySubtitle/vtt_generation.py
+PySubtitle.egg-info/PKG-INFO
+PySubtitle.egg-info/SOURCES.txt
+PySubtitle.egg-info/dependency_links.txt
+PySubtitle.egg-info/requires.txt
+PySubtitle.egg-info/top_level.txt
+example/example_usage.py
+tests/__init__.py
+tests/chunk0.wav
+tests/test.mp4
+tests/test_audio_extraction.py
+tests/test_speech_recognition.py
+tests/test_vtt_generation.py
