--- PySubtitle/.gitignore
@@ -1,4 +1,0 @@
-*.vtt
-.env
-.vscode
-.venv
