--- recognized.txt
@@ -1,29 +1,0 @@
-today we're going to be practicing you're speaking and conversational skills by having a conversation. 
- you and me what is it going to look like I have prepared a dialogue one line is for me and. 
- are there one is for you you will see them on the screen I'm going to say my line and you're going to read your line out. 
- on the screen and that's going to be your response and then vice versa the first line will be yours and the second mine this. 
- what size will give you already ready lines and answers to use in real life with other people and get your prepared. 
- I prepared a part of a dialogue between two friends or. 
- who haven't seen each other in a while and accidentally run into each other which is the story of my life I chose this type of. 
- station because honestly this is pretty much the conversation I've been having with people lately because it's true you never have the. 
- there's so many things to do and there's just. 
- four hours in a day which as it turns out is absolutely not enough so let's start I'm saying my line and then you will see. 
- online on the screen and said out loud as if you were answering to me oh hey have you been it's been a while. 
- oh I have so much to tell you are you in any kind of a hurry or how about grabbing coffee and catching up. 
- awesome let's go then who the thought would Cross. 
- like this. 
- does whole time but I've just been so busy lately so many things have been going on and I've just been doing what I can to. 
- and now we switch Lines the first one will be yours and the second mind so now you start. 
- hi Anna yeah. 
- I haven't seen you in ages so what's new with you. 
- sure sounds great I don't have to be anywhere. 
- I know right what a coincidence. 
- I know exactly what you mean it feels like life has become a little crazy. 
- okay that's it guys I hope you enjoyed this little practice video and found it useful and you know let me know in the comments. 
- if you can relate to such a situation bye. 
- what year it has I haven't seen you in ages so what's new with you oh I have so much to tell you are you in any kind of a. 
- or how about grabbing coffee and catching up sure sounds great I don't have to be anywhere awesome let's go then. 
- what a coincidence I have honestly been meaning to call you this whole time but I've just been. 
- busy leave me so many things have been going on and I've just been doing what I can to handle it all I know exactly what you mean. 
- it feels like life has become a little crazy lately. 
-
